+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKC2FSWPKR17AX8J97GRCVEF"
+        "01KKCE2SSA6X4S746DDE3RG6ZP"
     )
     r"""The unique identifier of the data source"""