+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKC2FSWB7MFJG9MMPJ9DH775"
+        "01KKCE2SS7JH8QKHRKF95CW425"
     )
     r"""The unique identifier of the data source"""