+++ packages/orq-rc/RELEASES.md
@@ -6459,3 +6459,13 @@ Based on:
 - [python v4.5.0-rc.48] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.48 - packages/orq-rc
+
+## 2026-03-10 17:55:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.750.0 (2.856.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.49] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.49 - packages/orq-rc